% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegimens.R
\name{createRegimens}
\alias{createRegimens}
\title{Create treatment regimens of a cohort}
\usage{
createRegimens(
  connectionDetails,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  cohortTable = cohortTable,
  rawEventTable = rawEventTable,
  regimenTable = regimenTable,
  regimenIngredientTable = regimenIngredientTable,
  vocabularyTable = vocabularyTable,
  drugClassificationIdInput = 21601387,
  cancerConceptId = 4115276,
  dateLagInput,
  generateVocabTable = TRUE,
  generateRawEvents = FALSE,
  sampleSize = 999999999999
)
}
\arguments{
\item{writeDatabaseSchema}{}

\item{rawEventTable}{}

\item{drugClassificationIdInput}{}

\item{dateLagInput}{.}

\item{generateVocabTable}{}

\item{generateRawEvents}{@return
SQL table in writeDatabaseSchema contains regimenIngredientTable.}

\item{sampleSize}{}

\item{Connection}{}

\item{CdmDatabaseSchema}{}

\item{CohortTable}{}

\item{regimenRepeats}{}
}
\description{
Creates treatment regimens from chosen classification code
}
