% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegimens.R
\name{createRegimens}
\alias{createRegimens}
\title{Create an oncology drug regimen table in a `writeDatabaseSchema` database}
\usage{
createRegimens(
  connectionDetails,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  cohortTable,
  rawEventTable,
  regimenTable,
  regimenIngredientTable,
  vocabularyTable,
  cancerConceptId = 4115276,
  dateLagInput = 30,
  generateVocabTable = FALSE,
  generateRawEvents = FALSE,
  keepSteroids = FALSE,
  useHemoncToPullDrugs = FALSE,
  writeToEpisodeTable = FALSE,
  writeToEpisodeEventTable = FALSE,
  episodeTypeConceptId = 32545,
  episodeEventTableConceptId = 1147094
)
}
\arguments{
\item{keepSteroids}{Boolean parameter if TRUE algorithm will look for steroids along other drugs}

\item{useHemoncToPullDrugs}{Boolean parameter if TRUE algorithm will use HemOnc vocabulary as a source of ingredients otherwise - internal csv}

\item{writeToEpisodeTable}{Boolean parameter if TRUE algorithm will delete form episode table with `episodeTypeConceptId` and insert `regimenIngredientTable` rows with `episodeTypeConceptId`}

\item{writeToEpisodeEventTable}{Boolean parameter if TRUE algorithm will delete form episode_event table with `episodeEventTableConceptId` and insert `regimenIngredientTable` rows with `episodeEventTableConceptId`}
}
\value{
This function does not return a value. It is called for its side effect of
creating a new SQL table called `regimenIngredientTable` in `writeDatabaseSchema`.
}
\description{
Creates treatment regimens
}
